+++ aworld/core/task.py
@@ -6,25 +6,25 @@ import traceback
 
 import uuid
 from typing import Union, Dict, Any, List
-from dataclasses import dataclass
+from dataclasses import dataclass, field
 from pydantic import BaseModel
 
 from aworld.core.agent.base import BaseAgent
-from aworld.config import ToolConfig, load_config, wipe_secret_info
-from aworld.core.common import Observation
+from aworld.core.common import Observation, ActionModel
 from aworld.core.envs.tool import Tool, ToolFactory
-from aworld.core.swarm import Swarm
+from aworld.core.agent.swarm import Swarm
 from aworld.logs.util import logger, color_log
 
 
 @dataclass
 class TaskModel:
-    name: str | None = uuid.uuid1().hex
-    input: Any | None = None
-    conf: Union[Dict[str, Any], BaseModel] | None = None
-    tools: List[Tool] | None = None
-    swarm: Swarm | None = None
-    agent: BaseAgent | None = None
+    name: str = uuid.uuid1().hex
+    input: Any = None
+    conf: Union[Dict[str, Any], BaseModel] = None
+    tools: List[Tool] = field(default_factory=list)
+    tool_names: List[str] = field(default_factory=list)
+    swarm: Swarm = None
+    agent: BaseAgent = None
 
 
 class Task(object):
@@ -35,7 +35,8 @@ class Task(object):
                  name: str = uuid.uuid1().hex,
                  input: Any = None,
                  conf: Union[Dict[str, Any], BaseModel] = {},
-                 tools: List[Tool] = None,
+                 tools: List[Tool] = [],
+                 tool_names: List[str] = [],
                  *args,
                  **kwargs):
         """Task instance init.
@@ -57,23 +58,37 @@ class Task(object):
             input = task.input
             conf = task.conf
             tools = task.tools
+            tool_names = task.tool_names
+            if tools is None:
+                tools = []
+            if tool_names is None:
+                tool_names = []
+
+        if not agent and not swarm:
+            raise ValueError("agent and swarm all is None.")
+        if agent and swarm:
+            raise ValueError("agent and swarm choose one only.")
+        if agent:
+            # uniform agent
+            swarm = Swarm(agent)
 
         if conf is None:
             conf = dict()
         if isinstance(conf, BaseModel):
             conf = conf.model_dump()
-        if not agent and not swarm:
-            raise ValueError("agent and swarm all is None.")
         check_input = conf.get("check_input", False)
         if check_input and not input:
             raise ValueError
 
-        self.agent = agent
         self.swarm = swarm
         self.input = input
         self.name = name
         self.conf = conf
         self.tools = {tool.name(): tool for tool in tools} if tools else {}
+        tool_names.extend(self.tools.keys())
+        # lazy load
+        self.tool_names = tool_names
+
         self.daemon_target = kwargs.pop('daemon_target', None)
         self._use_demon = False if not conf else conf.get('use_demon', False)
         self._exception = None
@@ -120,11 +135,7 @@ class Task(object):
         else:
             observation = Observation(content=self.input)
 
-        if self.agent:
-            return self._agent_process(observation, info)
-        elif self.swarm:
-            # example now
-            return self._swarm_process(observation, info)
+        return self._swarm_process(observation, info)
 
     def _swarm_process(self,
                        observation: Observation,
@@ -132,7 +143,7 @@ class Task(object):
         start = time.time()
 
         input = observation.content
-        self.swarm.reset(self.tools)
+        self.swarm.reset(self.tool_names)
         for agent in self.swarm.agents.values():
             agent.reset({"task": input})
 
@@ -142,7 +153,7 @@ class Task(object):
         try:
             while step < max_steps:
                 # Loose protocol
-                result_dict = self.swarm.process(observation=observation, info=info)
+                result_dict = self._process(observation=observation, info=info)
                 results.append(result_dict)
 
                 swarm_resp = result_dict.get("response")
@@ -185,85 +196,158 @@ class Task(object):
                     "success": False,
                     "total_time": (time.time() - start)}
 
-    def _agent_process(self,
-                       observation: Observation,
-                       info: Dict[str, Any]) -> Dict[str, Any]:
-        agent = self.agent
-        agent.reset({"task": observation.content})
+    def is_agent(self, policy: ActionModel):
+        return policy.tool_name is None and policy.action_name is None
+
+    def _process(self, observation, info) -> Dict[str, Any]:
+        """Multi-agent general process workflow.
+
+        NOTE: Use the agent‘s finished state to control the loop, so the agent must carefully set finished state.
+        Args:
+            observation: Observation based on env
+            info: Extend info by env
+        """
+        if not self.swarm.initialized:
+            raise RuntimeError("swarm needs to use `reset` to init first.")
+
+        start = time.time()
         step = 0
         max_steps = self.conf.get("max_steps", 100)
-        excep = None
-        start = time.time()
+        # use entry agent every time
+        policy: List[ActionModel] = self.swarm.cur_agent.policy(observation, info)
+        if not policy:
+            logger.warning(f"current agent {self.swarm.cur_agent.name()} no policy to use.")
+            return {"msg": f"current agent {self.swarm.cur_agent.name()} no policy to use.",
+                    "steps": step,
+                    "success": False,
+                    "time_cost": (time.time() - start)}
+
+        if self.tools is None:
+            self.tools = {}
+
+        msg = None
+        response = None
+        return_entry = False
         try:
             while step < max_steps:
                 terminated = False
-                # Get policy from agent, policy must is an action list or None
-                policy = agent.policy(observation=observation, info=info)
-
-                if not policy:
-                    logger.warning(f"{observation} can not get the valid policy, exit task!")
-                    excep = "invalid policy"
-                    break
-
-                # group action by tool name
-                tool_mapping = dict()
-                # Directly use or use tools after creation.
-                for act in policy:
-                    if not self.tools or (self.tools and act.tool_name not in self.tools):
-                        # only use default config in module or XXConfig.
-                        conf = load_config(f"{act.tool_name}.yaml")
-                        if not conf:
-                            conf = ToolConfig()
-                        tool = ToolFactory(act.tool_name, conf=conf)
-                        logger.info(f"Dynamic load config from {act.tool_name}.yaml, "
-                                    f"conf is: {wipe_secret_info(conf, ['api_key'])}")
-                        tool.reset()
-                        tool_mapping[act.tool_name] = []
-                        self.tools[act.tool_name] = tool
-                    if act.tool_name not in tool_mapping:
-                        tool_mapping[act.tool_name] = []
-                    tool_mapping[act.tool_name].append(act)
-
-                for tool_name, action in tool_mapping.items():
-                    # Execute action using browser tool and unpack all return values
-                    observation, reward, terminated, _, info = self.tools[tool_name].step(action)
-                    # need merge?
-
-                    logger.info(f'{action} state: {observation}; reward: {reward}')
-                    # Check if there's an exception in info
-                    if info.get("exception"):
-                        color_log(f"Step {step} failed with exception: {info['exception']}")
-                        excep = info.get("exception")
-
-                    step += 1
-                    logger.info(f"step: {step} finished by tool action.")
-                    # Check if task should end (either terminated or truncated)
-                    if agent.finished:
-                        logger.info("agent or tools finished")
+                if self.is_agent(policy[0]):
+                    # only one agent, and get agent from policy
+                    policy_for_agent = policy[0]
+                    cur_agent: BaseAgent = self.swarm.agents.get(policy_for_agent.agent_name)
+                    if not cur_agent:
+                        raise RuntimeError(f"Can not find {policy_for_agent.agent_name} agent in swarm.")
+                    if cur_agent.name() == self.swarm.communicate_agent.name():
+                        # Current agent is entrance agent, means need to exit to the outer loop
+                        logger.warning("Exit to the outer loop")
+                        return_entry = True
                         break
 
-                if terminated:
-                    logger.info(f"{self.name} task finished")
+                    if self.swarm.cur_agent.handoffs and policy_for_agent.agent_name not in self.swarm.cur_agent.handoffs:
+                        # Unable to hand off, exit to the outer loop
+                        return {"msg": f"Can not handoffs {policy_for_agent.agent_name} agent "
+                                       f"by {cur_agent.name()} agent.",
+                                "response": policy[0].policy_info if policy else "",
+                                "steps": step,
+                                "success": False}
+                    # Check if current agent done
+                    if cur_agent.finished:
+                        cur_agent._finished = False
+                        logger.info(f"{cur_agent.name()} agent be be handed off, so finished state reset to False.")
+
+                    observation = Observation(content=policy_for_agent.policy_info)
+                    agent_policy = cur_agent.policy(observation, info=info)
+                    if not agent_policy:
+                        logger.warning(
+                            f"{observation} can not get the valid policy in {policy_for_agent.agent_name}, exit task!")
+                        msg = f"{policy_for_agent.agent_name} invalid policy"
+                        break
+                    self.swarm.cur_agent = cur_agent
+                    policy = agent_policy
+                    # clear observation
+                    observation = None
+                else:
+                    # group action by tool name
+                    tool_mapping = dict()
+                    # Directly use or use tools after creation.
+                    for act in policy:
+                        if not self.tools or (self.tools and act.tool_name not in self.tools):
+                            # dynamic only use default config in module.
+                            tool = ToolFactory(act.tool_name)
+                            tool.reset()
+                            tool_mapping[act.tool_name] = []
+                            self.tools[act.tool_name] = tool
+                        if act.tool_name not in tool_mapping:
+                            tool_mapping[act.tool_name] = []
+                        tool_mapping[act.tool_name].append(act)
+
+                    for tool_name, action in tool_mapping.items():
+                        # Execute action using browser tool and unpack all return values
+                        observation, reward, terminated, _, info = self.tools[tool_name].step(action)
+
+                        logger.info(f'{action} state: {observation}; reward: {reward}')
+                        # Check if there's an exception in info
+                        if info.get("exception"):
+                            color_log(f"Step {step} failed with exception: {info['exception']}")
+                            msg = info.get("exception")
+                        logger.info(f"step: {step} finished by tool action.")
+
+                    # The tool results give itself, exit; give to other agents, continue
+                    tmp_name = policy[0].agent_name
+                    if self.swarm.cur_agent.name() == self.swarm.communicate_agent.name() and (
+                            len(self.swarm.agents) == 1 or tmp_name is None or self.swarm.cur_agent.name() == tmp_name):
+                        return_entry = True
+                        break
+                    elif policy[0].agent_name:
+                        policy_for_agent = policy[0]
+                        cur_agent: BaseAgent = self.swarm.agents.get(policy_for_agent.agent_name)
+                        if not cur_agent:
+                            raise RuntimeError(f"Can not find {policy_for_agent.agent_name} agent in swarm.")
+                        if self.swarm.cur_agent.handoffs and policy_for_agent.agent_name not in self.swarm.cur_agent.handoffs:
+                            # Unable to hand off, exit to the outer loop
+                            return {"msg": f"Can not handoffs {policy_for_agent.agent_name} agent "
+                                           f"by {cur_agent.name()} agent.",
+                                    "response": policy[0].policy_info if policy else "",
+                                    "steps": step,
+                                    "success": False}
+                        # Check if current agent done
+                        if cur_agent.finished:
+                            cur_agent._finished = False
+                            logger.info(f"{cur_agent.name()} agent be be handed off, so finished state reset to False.")
+
+                step += 1
+                if terminated and self.swarm.cur_agent.finished:
+                    logger.info("swarm finished")
                     break
 
-            time_cost = time.time() - start
+                if observation:
+                    if cur_agent is None:
+                        cur_agent = self.swarm.cur_agent
+                    policy = cur_agent.policy(observation, info)
+
+            if policy:
+                response = policy[0].policy_info
+
+            # All agents or tools have completed their tasks
+            if all(agent.finished for _, agent in self.swarm.agents.items()) or (all(
+                    tool.finished for _, tool in self.tools.items()) and len(self.swarm.agents) == 1):
+                logger.info("entry agent finished, swarm process finished.")
+                self.finished = True
+
+            if return_entry:
+                # Return to the entrance, reset current agent finished state
+                self.swarm.cur_agent._finished = False
             return {"steps": step,
-                    "msg": excep,
-                    "success": True if not excep else False,
-                    "total_time": time_cost}
+                    "response": response,
+                    "observation": observation,
+                    "msg": msg,
+                    "success": True if not msg else False}
         except Exception as e:
             logger.error(f"Task execution failed with error: {str(e)}\n{traceback.format_exc()}")
-            time_cost = time.time() - start
             return {
                 "msg": str(e),
+                "response": "",
                 "traceback": traceback.format_exc(),
                 "steps": step,
-                "success": False,
-                "total_time": time_cost
+                "success": False
             }
-        finally:
-            # Cleanup if not keeping open
-            if self.tools:
-                for _, tool in self.tools.items():
-                    if not tool.dict_conf.get("keep_open", False):
-                        tool.close()
